retry if get error while reading ByteStream (#395)
We can hit a failure of the `x.unwrap()` in get_object_impl().  This is probably due to the connection to S3 dying after we get the header but while in the middle of getting the payload.  As noted in the XXX comment, we need to retry when we hit this error.

This PR makes us retry when we hit this error.  I realized that there was quite a bit of cleanup available related to the retry code, which made it possible to use the generic `object_access::retry()` to retry this error as well.

- since the retry() callback can return Ok(any type), we can move all the logic inside it (for delete_objects() and get_object_impl())
- make retry() callback return Err(OAError) and add Other variant to OAError, so that any type of error can be returned (not just RusotoError) [not sure this is really necessary since we could probably just create a RusotoError::HttpDispatch from the S3Error in delete_objects()]

side cleanup:
- make `object_access::delete_objects()` do the chunking for us
- make chunk size configurable
- since retry_impl() already handles not retrying RusotoError::Service, we don't need to do this in get_object_impl().  That was the only callback that returned (false, _), so the callback doesn't need to return the "can retry" boolean.