agent resume: limit writes to those before the next recovered block (#396)
When the agent is resuming (i.e. agent restarted, kernel reconnects and resumes a txg), the kernel re-issues any outstanding writes to the agent. The agent finds any objects that were written before it crashed ("recovered objects"), and only creates new objects ("gap objects") for blocks that were not already stored to objects. i.e. If there's a gap in the objects, we need to fill it in, based on the writes that the kernel gave us.

The problem is that the agent was putting as many writes as possible into the "gap object", but it needs to limit this to only blocks that are before the next "recovered object".