fix resuming just after writing super object (#400)
If the agent dies during the "end txg" command, after writing the super object but before the kernel gets the "end txg done" message, then when we try to resume, the kernel will try to resume the txg, but the agent has already completed this txg, leading the agent to fail an assertion.

This PR fixes this by having the kernel detect this situation and skip the "resume txg" command, but re-send the "end txg". The agent is changed to handle the duplicate "end txg" by re-sending the "end txg done", which is what the kernel expects.